% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProgram.R
\name{dist.JSD_2m_fast}
\alias{dist.JSD_2m_fast}
\title{scProgram}
\usage{
dist.JSD_2m_fast(inMatrix1, inMatrix2)
}
\description{
inMatrix1
}
\examples{
dist.JSD_2m_fast()
GetFeatures()
HeatFeatures()
GetProgram()
}
\keyword{scProgram}
